+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 11.4.0
+
+* `[system]` Added method `ValidatePerms` for permissions validation
+* `[system]` Increased code coverage (98.0% → 98.8%)
+
 ### 11.3.1
 
 * `[initsystem]` Fixed stubs for Windows